+++ app/Console/Commands/AutoStatisticsNodeDailyTraffic.php
@@ -3,16 +3,15 @@
 namespace App\Console\Commands;
 
 use Illuminate\Console\Command;
-use App\Http\Models\User;
 use App\Http\Models\SsNode;
+use App\Http\Models\SsNodeTrafficDaily;
 use App\Http\Models\UserTrafficLog;
-use App\Http\Models\UserTrafficHourly;
 use Log;
 
-class AutoStatisticsUserHourlyTrafficJob extends Command
+class AutoStatisticsNodeDailyTraffic extends Command
 {
-    protected $signature = 'autoStatisticsUserHourlyTrafficJob';
-    protected $description = '自动统计用户每小时流量';
+    protected $signature = 'autoStatisticsNodeDailyTraffic';
+    protected $description = '自动统计节点每日流量';
 
     public function __construct()
     {
@@ -21,39 +20,27 @@ class AutoStatisticsUserHourlyTrafficJob extends Command
 
     public function handle()
     {
-        $userList = User::query()->where('status', '>=', 0)->where('enable', 1)->get();
-        foreach ($userList as $user) {
-            // 统计一次所有节点的总和
-            $this->statisticsByNode($user->id);
-
-            // 统计每个节点产生的流量
-            $nodeList = SsNode::query()->where('status', 1)->orderBy('id', 'asc')->get();
-            foreach ($nodeList as $node) {
-                $this->statisticsByNode($user->id, $node->id);
-            }
+        $nodeList = SsNode::query()->where('status', 1)->orderBy('id', 'asc')->get();
+        foreach ($nodeList as $node) {
+            $this->statisticsByNode($node->id);
         }
 
         Log::info('定时任务：' . $this->description);
     }
 
-    private function statisticsByNode($user_id, $node_id = 0)
+    private function statisticsByNode($node_id)
     {
-        $start_time = strtotime(date('Y-m-d H:i:s', strtotime("-1 hour")));
-        $end_time = time();
+        $start_time = strtotime(date('Y-m-d 00:00:00', strtotime("-1 day")));
+        $end_time = strtotime(date('Y-m-d 23:59:59', strtotime("-1 day")));
 
-        $query = UserTrafficLog::query()->where('user_id', $user_id)->whereBetween('log_time', [$start_time, $end_time]);
-
-        if ($node_id) {
-            $query->where('node_id', $node_id);
-        }
+        $query = UserTrafficLog::query()->where('node_id', $node_id)->whereBetween('log_time', [$start_time, $end_time]);
 
         $u = $query->sum('u');
         $d = $query->sum('d');
         $total = $u + $d;
         $traffic = $this->flowAutoShow($total);
 
-        $obj = new UserTrafficHourly();
-        $obj->user_id = $user_id;
+        $obj = new SsNodeTrafficDaily();
         $obj->node_id = $node_id;
         $obj->u = $u;
         $obj->d = $d;
