+++ app/Console/Commands/AutoStatisticsNodeHourlyTraffic.php
@@ -4,14 +4,14 @@ namespace App\Console\Commands;
 
 use Illuminate\Console\Command;
 use App\Http\Models\SsNode;
-use App\Http\Models\SsNodeTrafficDaily;
+use App\Http\Models\SsNodeTrafficHourly;
 use App\Http\Models\UserTrafficLog;
 use Log;
 
-class AutoStatisticsNodeDailyTrafficJob extends Command
+class AutoStatisticsNodeHourlyTraffic extends Command
 {
-    protected $signature = 'autoStatisticsNodeDailyTrafficJob';
-    protected $description = '自动统计节点每日流量';
+    protected $signature = 'autoStatisticsNodeHourlyTraffic';
+    protected $description = '自动统计节点每小时流量';
 
     public function __construct()
     {
@@ -30,8 +30,8 @@ class AutoStatisticsNodeDailyTrafficJob extends Command
 
     private function statisticsByNode($node_id)
     {
-        $start_time = strtotime(date('Y-m-d 00:00:00', strtotime("-1 day")));
-        $end_time = strtotime(date('Y-m-d 23:59:59', strtotime("-1 day")));
+        $start_time = strtotime(date('Y-m-d H:i:s', strtotime("-1 hour")));
+        $end_time = time();
 
         $query = UserTrafficLog::query()->where('node_id', $node_id)->whereBetween('log_time', [$start_time, $end_time]);
 
@@ -40,7 +40,7 @@ class AutoStatisticsNodeDailyTrafficJob extends Command
         $total = $u + $d;
         $traffic = $this->flowAutoShow($total);
 
-        $obj = new SsNodeTrafficDaily();
+        $obj = new SsNodeTrafficHourly();
         $obj->node_id = $node_id;
         $obj->u = $u;
         $obj->d = $d;
