+++ app/Console/Kernel.php
@@ -13,28 +13,20 @@ class Kernel extends ConsoleKernel
      * @var array
      */
     protected $commands = [
-        \App\Console\Commands\AutoBanSubscribeJob::class,
-        \App\Console\Commands\AutoBanUserJob::class,
-        \App\Console\Commands\AutoCheckNodeStatusJob::class,
-        \App\Console\Commands\AutoClearLogJob::class,
-        \App\Console\Commands\AutoCloseOrderJob::class,
-        \App\Console\Commands\AutoDecGoodsTrafficJob::class,
-        \App\Console\Commands\AutoDisableExpireUserJob::class,
-        \App\Console\Commands\AutoDisableUserJob::class,
-        \App\Console\Commands\AutoExpireCouponJob::class,
-        \App\Console\Commands\AutoExpireInviteJob::class,
-        \App\Console\Commands\AutoReleasePortJob::class,
-        \App\Console\Commands\AutoRegetPortJob::class,
-        \App\Console\Commands\AutoRemoveDisabledUserLabelsJob::class,
-        \App\Console\Commands\AutoReopenUserJob::class,
-        \App\Console\Commands\AutoResetUserTrafficJob::class,
-        \App\Console\Commands\AutoStatisticsNodeDailyTrafficJob::class,
-        \App\Console\Commands\AutoStatisticsNodeHourlyTrafficJob::class,
-        \App\Console\Commands\AutoStatisticsUserDailyTrafficJob::class,
-        \App\Console\Commands\AutoStatisticsUserHourlyTrafficJob::class,
-        \App\Console\Commands\UserExpireWarningJob::class,
-        \App\Console\Commands\UserTrafficAbnormalWarningJob::class,
-        \App\Console\Commands\UserTrafficWarningJob::class,
+        \App\Console\Commands\AutoJob::class,
+        \App\Console\Commands\AutoClearLog::class,
+
+        \App\Console\Commands\AutoDecGoodsTraffic::class,
+        \App\Console\Commands\AutoResetUserTraffic::class,
+
+        \App\Console\Commands\AutoStatisticsNodeDailyTraffic::class,
+        \App\Console\Commands\AutoStatisticsNodeHourlyTraffic::class,
+        \App\Console\Commands\AutoStatisticsUserDailyTraffic::class,
+        \App\Console\Commands\AutoStatisticsUserHourlyTraffic::class,
+
+        \App\Console\Commands\UserTrafficAbnormalAutoWarning::class,
+        \App\Console\Commands\UserExpireAutoWarning::class,
+        \App\Console\Commands\UserTrafficAutoWarning::class,
     ];
 
     /**
@@ -46,28 +38,20 @@ class Kernel extends ConsoleKernel
      */
     protected function schedule(Schedule $schedule)
     {
-        $schedule->command('autoReopenUserJob')->everyMinute();
-        $schedule->command('autoBanSubscribeJob')->everyThirtyMinutes();
-        $schedule->command('autoBanUserJob')->everyTenMinutes();
-        $schedule->command('autoCheckNodeStatusJob')->everyMinute();
-        $schedule->command('autoClearLogJob')->everyThirtyMinutes();
-        $schedule->command('autoCloseOrderJob')->everyMinute();
-        $schedule->command('autoDecGoodsTrafficJob')->everyTenMinutes();
-        $schedule->command('autoDisableExpireUserJob')->everyMinute();
-        $schedule->command('autoDisableUserJob')->everyMinute();
-        $schedule->command('autoExpireCouponJob')->everyThirtyMinutes();
-        $schedule->command('autoExpireInviteJob')->everyThirtyMinutes();
-        $schedule->command('autoRemoveDisabledUserLabelsJob')->everyMinute();
-        $schedule->command('autoReleasePortJob')->everyMinute();
-        $schedule->command('autoRegetPortJob')->everyMinute();
-        $schedule->command('autoResetUserTrafficJob')->everyFiveMinutes();
-        $schedule->command('autoStatisticsNodeDailyTrafficJob')->dailyAt('23:55');
-        $schedule->command('autoStatisticsNodeHourlyTrafficJob')->hourly();
-        $schedule->command('autoStatisticsUserDailyTrafficJob')->dailyAt('23:50');
-        $schedule->command('autoStatisticsUserHourlyTrafficJob')->hourly();
-        $schedule->command('userExpireWarningJob')->daily();
-        $schedule->command('userTrafficAbnormalWarningJob')->everyThirtyMinutes();
-        $schedule->command('userTrafficWarningJob')->daily();
+        $schedule->command('autoJob')->everyMinute();
+        $schedule->command('autoClearLog')->everyThirtyMinutes();
+
+        $schedule->command('autoDecGoodsTraffic')->everyTenMinutes();
+        $schedule->command('autoResetUserTraffic')->everyFiveMinutes();
+
+        $schedule->command('autoStatisticsNodeDailyTraffic')->dailyAt('23:55');
+        $schedule->command('autoStatisticsNodeHourlyTraffic')->hourly();
+        $schedule->command('autoStatisticsUserDailyTraffic')->dailyAt('23:50');
+        $schedule->command('autoStatisticsUserHourlyTraffic')->hourly();
+
+        $schedule->command('userTrafficAbnormalAutoWarning')->everyThirtyMinutes();
+        $schedule->command('userExpireAutoWarning')->daily();
+        $schedule->command('userTrafficAutoWarning')->daily();
     }
 
     /**
