+++ app/Http/Middleware/VerifyCsrfToken.php
@@ -12,6 +12,6 @@ class VerifyCsrfToken extends BaseVerifier
      * @var array
      */
     protected $except = [
-        "payment/*", "mu/v2/*"
+        "payment/*", "Muv2/*"
     ];
 }
