+++ app/Http/V2ray/EmptyClass.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace App\Http\V2ray;
+namespace App\Http\V2Ray;
 
 
 class EmptyClass
