+++ routes/api.php
@@ -4,5 +4,7 @@ Route::group(['namespace' => 'Api'], function () {
     Route::any('yzy/create', 'YzyController@create');
     Route::resource('yzy', 'YzyController');
 
+    // 定制客户端
     Route::get('login', 'LoginController@login');
+
 });
